Change spatial to RHR and column major notation
	Despite a seemingly functional camera demo, the matrix math was
actually really janky. This commit makes the spatial module more
consistent.

	The first change is that OpenGL uses right hand coordinates by
default, the camera demo has been updated to this fact.

	The second change is that OpenGL and NAlgebra both use
column-major notation for their matricies, this leads to the unintuitive
order in which transformation matricies need to be applied. While we
need to translate spatials before we rotate them for objects to move as
we expect, the transformation matrix of our spatial must be multiplied
by our translation FIRST.

	One small consequence of all this is that the camera's model
space matrix can no longer be used directly as the world space matrix,
we now have to get the inverse matrix.